fix short type range, migrate to int

+++ excel-exporter/src/main/java/pl/com/softproject/utils/excelexporter/ExcelExporter.java
@@ -40,7 +40,7 @@ public class ExcelExporter {
     protected Sheet sheet;
     protected Workbook wb;
     protected Row header;
-    protected short currentRowNumber;
+    protected int currentRowNumber;
     protected int currentColumnNumber;
 
     protected DataFormat format;
